requirements.txt
google-genai>=1.0
openai>=1.0
python-dotenv>=1.0
rich>=13.0
